+++ src/test/java/dev/group4/ItemServiceTests.java
@@ -1,9 +1,10 @@
-package dev.group4.services;
+package dev.group4;
 
 import dev.group4.entities.Item;
 import dev.group4.entities.Potluck;
 import dev.group4.entities.StatusType;
 import dev.group4.repos.ItemRepo;
+import dev.group4.services.ItemServiceImpl;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
